refactor: AuctionLotsForYou section
